This PR works on various issues around GC, trying to find
leaks/incorrect bits etc.

There are a lot of bits here that change code to slowly get clang-tidy
become useful, which also includes reducing undefined behavior etc.

The major changes are:
 - add a copying GC that's backed per-object by malloc: DebugCopyingGC
 - add `IsMarkedValid` for more aggressive assertion checking
 - add logging levels
- refactor helpers to write pointers, to make more explicit what the
GC-related invariants are
 - make sure object hash is stable across GCs
- rework object creation to avoid undefined behavior (pass total object
size into constructor)

+++ src/compiler/Lexer.cpp
@@ -25,6 +25,7 @@
  */
 #include <cassert>
 #include <cctype>
+#include <cstdint>
 #include <cstring>
 #include <istream>
 #include <string>
@@ -38,7 +39,7 @@ Lexer::Lexer(istream &file) : infile(file), peekDone(false) {}
 #define _BC (buf[state.bufp])
 #define EOB (state.bufp >= buf.length())
 
-size_t Lexer::fillBuffer() {
+int64_t Lexer::fillBuffer() {
     if (!infile.good()) { // file stream
         return -1;
     }
